+++ R/code-pow.R
@@ -20,8 +20,14 @@ code_gen_pow <-  function(fun, pars, par.types) {
     par.types=character() && all(. %in% PAR.IVARY)
   )
   name <- FUN.NAMES[fun]
+  # Test is not perfect, ideally it would match what USE_POWL_IN_R_POW
+  # is set by (from src/gnuwin32/fixed/h/config.h as of ~2024-01)
+  win <- identical(.Platform[['OS.type']], "windows")
+  x64 <- identical(.Platform[['r_arch']], "x64")
+  pow.fun <- if(win && x64) "powl" else "pow"
+
   defn <- sprintf(
-    bin_op_vec_rec, name, toString(CF.ARGS.BASE), "pow", ",",
+    bin_op_vec_rec, name, toString(CF.ARGS.BASE), pow.fun, ",",
     IX[['I.STAT']], IX[['STAT.RECYCLE']]
   )
   code_res(defn=defn, name=name, headers="<math.h>")